Renamed invokeBatonHandler file name and update test import
